Task 7:
Create a Student class in the same project. In the Student class:
a. We have private data members for the student's first name, last name, student ID, age, and GPA.
b. There is a default constructor that initializes the data members with default values (empty strings, 0, and 0.0).
c. We also have a convenience constructor that allows you to sell all the data members when creating a Student object.
d. Getter and setter methods (not shown in the code) should be added for accessing and modifying the data members.
e. The toString method is overriden to provide a human-readable representation of the Student object's information.