Task 8:
Create a StudentTest class in the same project. In the StudentTest class: a. We create two Student objects, one using the default constructor named student1 and another using the convenience constructor named student2. b. Create a method printStudent. The printStudent method takes a Student object as a parameter and uses the toString method of the Student Class to print the student's information in a human-readable format. toString should not be called explicitly
c. We then csll the printStudent method from the main method to print the information for each student.